We'd like to have specific labels for relationships between GND entities and thus replace the generic :RELA-label by the GND codes for relations defined here: http://www.dnb.de/SharedDocs/Downloads/DE/DNB/standardisierung/inhaltserschliessung/gndCodes.pdf?__blob=publicationFile. Using these codes as labels, we can remove the id-property which formerly has contained the codes.
This approach requires also to restructure the csv files. We have to define for all codes a separate --relationships:code argument. Thus we had to refactor the mechanisms responsible writing the content to the csv files. By doing so, we made the same changes to the nodes, too, which don't have a LABEL column anymore and have to be called by providing a respective --nodes:LABEL argument.