Extends JsonDecoder in order to support the decoding of more complex (i.e. deeply nested) JSON objects
JSON elements like objects, arrays, literals or "special literals" (e.g. numbers) are now represented as Java objects instead of mere numbers. This allows to save key values along with the type of the respective element. A "path" list tracks the current nesting of these elements during the parsing of the JSON documents. Thus it is now possible to parse e.g. objects inside an array of parent objects.