Allow inclusion / exlusion of language tag or type annotation respectively
Provide two parameters `keepLanguageTags` and `keepTypeAnnotations` for controlling if language tags or data type annotations should be appended to output literal strings if they are part of the original literal. The annotations / tags are separated by two hash characters ('#') for enabling the extraction of the postfixes in the Morph file.

Signed-off-by: Sebastian Schüpbach <sebastian.schuepbach@unibas.ch>